requirements.txt
pykevoplus
paho-mqtt
